DriveStraight and Pivot Fully Tested
Tested DriveStraight and PivotCommand, and they both work. Disabled
commands if the robot was disabled earlier than expected. Changed ini
refresh to be a bit universal so the PIDFacs are not public variables in
robot model. Fixed PivotCommand so it's set continuous. Absolute
position works (haven't tested delta angle but prob works).

~ Lynn, Alisha